+++ source/tm/move/max_moves_per_pokemon.hpp
@@ -27,17 +27,17 @@ namespace technicalmachine {
 using namespace bounded::literal;
 
 constexpr auto max_moves_per_pokemon = 4_bi;
-using RegularMoveSize = bounded::integer<1, static_cast<intmax_t>(max_moves_per_pokemon)>;
+using RegularMoveSize = bounded::integer<1, static_cast<int>(max_moves_per_pokemon)>;
 using RegularMoveIndex = bounded::checked_integer<
 	0,
-	static_cast<intmax_t>(std::numeric_limits<RegularMoveSize>::max() - 1_bi)
+	static_cast<int>(std::numeric_limits<RegularMoveSize>::max() - 1_bi)
 >;
 
 constexpr auto number_of_weird_moves = 2_bi; // Pass, Struggle
 using SharedMoveSize = decltype(std::declval<TeamSize>() + number_of_weird_moves);
 using SharedMoveIndex = bounded::checked_integer<
 	0,
-	static_cast<intmax_t>(std::numeric_limits<SharedMoveSize>::max() - 1_bi)
+	static_cast<int>(std::numeric_limits<SharedMoveSize>::max() - 1_bi)
 >;
 
 using MoveSize = decltype(std::declval<RegularMoveSize>() + std::declval<SharedMoveSize>());
